[vfs] Attempt to short-circuit directory enumeration
In case-insensitive filesystems, canonizing a given name to the one in the filesystem is expensive if we need to enumerate directories. Attempt to compute an answer earlier by generating a full realpath, which should succeed in the absence of symbolic links.

https://github.com/JetBrains/intellij-community/pull/2856

(cherry picked from commit 344bd30cb8d04bf23905e6ab92731db3f66ae193)

Bug: 363903536
GitOrigin-RevId: 05a614b009f25acf4f764fbdde6db68e53ead7a3
Change-Id: I20083ccc2d277cf5b693ca07e1de742975503993