[vfs] Attempt to short-circuit directory enumeration
In case-insensitive filesystems, canonizing a given name to the one in the filesystem is expensive if we need to enumerate directories. Attempt to compute an answer earlier by generating a full realpath, which should succeed in the absence of symbolic links.

https://github.com/JetBrains/intellij-community/pull/2856


(cherry picked from commit 05a614b009f25acf4f764fbdde6db68e53ead7a3)

IJ-CR-146719

GitOrigin-RevId: 869546be6e79b6ee904250c11cf5775ad794054e