IJPL-160019 Fix restoring tree paths on workspace switch
This regression is caused by the fact that CachedTreePresentation
doesn't work well yet with trees that are already loaded.
In this particular case, though, it's not desirable to use
presentation caching because the cached presentation
may not match the actual tree at all.

So let's introduce an internal method to skip presentation caching
and use it there.

GitOrigin-RevId: 1d756971f87b4a2803d3e05b9ec41de7b7a76bb6