[plugin model] add paths from required modules to the classpath of the main classloader (IJPL-148565)
Usually required plugin modules are packed to the main plugin JAR. However, to allow running the frontend process from the monolith IDE, some modules may be put to separate JARs to ensure that the frontend process doesn't include unrelated classes in the classpath. In that case, JAR files for additional required modules must be added to the classpath of the main classloader, otherwise classes from such additional modules won't be found.

GitOrigin-RevId: 28909e8ae9d27dce3d4e28a81d8180eefdb4436c