IJPL-173842 gitlabci: basic code-insight support for "stage" symbol
it is a squash commit of several commits. It was made to avoid code appearance in the community part of the repository and because the gitlab-plugin layout (structure of the related modules, files, etc.) was changed. The messages of the squashed commits are:
- IJPL-84197 GitLab CI: add navigation from `stage` usage to its declaration
- IJPL-84197 GitLab CI: bidirectional navigation refs<->declarations for "stages" using Symbol API
- IJPL-84197 GitLab CI: completion for "stages" via Symbol API
- IJPL-84197 GitLab CI: rename "stages" in "in-place" mode via Symbol API considering "broken" ranges
"broken" ranges - ranges, that are passed into Pointer restoration function (see `GitLabCiStageSymbol.createPointer`) from IDE at the moment when Enter is pressed in the end of Rename action in "in-place" mode. "end" values of such ranges are equal to "start" values, and this looks like not expected behavior. And hacks are needed to consider these "broken" ranges.
- IJPL-84197 GitLab CI: support quoted strings for "stage"-values as well as plain text
- IJPL-84197 GitLab CI: fix problem: "stage declaration is not highlighted when cursor is on usage"
plus some refactoring
- IJPL-84197 GitLab CI: turn on registry flag
for simpler future tests running
- IJPL-84197 GitLab CI: tests for `stage` symbol editor support
- IJPL-84197 GitLab CI: move code to gitlab-ultimate module

GitOrigin-RevId: f49ea2717bfa4764abefee5ef81a9e474690754a