[platform] avoid read-only status dialog for project configuration files
These files are internal and should be handled by state store.
Such files can be either versioned (added to some version control) or unversioned. In the case of versioned files, the corresponding VCS know how to make it writable. In the case of unversiond, the default file system writable operation is executed.

GitOrigin-RevId: 712a565c94348bdb31643ed6b244672fd75cc71e