coroutines: ignore ControlFlowException in ContextRunnable
It could happen that ContextRunnable is run on JDK worker that
does not have special handling for CancellationExceptions
and logs them as errors "Exception in thread". However,
ControlFlowException is not a real error and can be safely ignored.
Since we can't fix this on JDK side, and we're not going to force
all runnables being run on intellij workers, we're ok to ignore the
exception right in ContextRunnable.

GitOrigin-RevId: a99e8f50ea89b6685c4205468a8b92577e1f1611