Add K2 debugger evaluation test with compose compiler plugin
We have to test whether the compiler plugins correctly work or not for
debugger evaluation. For example, When the debugger evaluates
`Foo(::getString)` for the following code:
```
@Composable
fun Foo(child: @Composable () -> Unit) {
    child()
}

fun getString() = "empty"

fun main() {
..
//Breakpoint!
}
```
it must report a diagnostic error triggered by the compose compiler
plugin checker because it passes a non-composable function for a
parameter with a `@Composable` annotation.

This commit adds a debugger evaluation test to check the above example
reports a diagnostic error correctly.

In addition, compiling the above KT file should work without an error.
If we run compose compiler plugin IrGenerationExtension multiple times,
it will have some errors from IR generation for `Foo()`.

To test the compose compiler plugin registration on IDE and its
functionality for the debugger evaluation, this commit updates
KotlinDescriptorTestCase to set the test module with the compiler
plugin.

This commit is the first part of testing debugger evaluation i.e.,
IrGenerationExtensions of various compiler plugins for K2 CodeGen API on
IDE.

KT-72347