IJPL-157713 Do not show splash above all windows
Splash was above all windows to make it show on top of the IDE until editor or tool windows initialized. But there was a problem that it also stayed on top of all other windows in the OS.
There was an approach to close the splash on IDE deactivation, but it was not working stable enough on Windows.
So, we have to not show the splash on top to not distract Windows users.

GitOrigin-RevId: cdf5cb30aee5cf08a384a81c6633e750ec658919