IJPL-56066 Drop MigLayout in ToolWindowHeader
Replace with GroupLayout, as we need something that
lets us control how the components grow.

Unfortunately, GroupLayout isn't very friendly about
adding and removing components dynamically,
so we have to rebuild the whole thing every time
we need to add or remove something, otherwise
the layout often ends up broken.
In practice, this means re-adding the toolbar
when the sideComponent is set. This is a performance
problem, but a very minor one, as sideComponent
is almost never used in practice.

GitOrigin-RevId: eeb7ec752879fbf7dc02934ad263d277227735f3