IJPL-45210 Add the editor.max.search.selection.length key
To avoid freezes when the user accidentaly Ctrl+A, Ctrl+F a huge file,
set a reasonable limit on the text length when the search/replace
fields are updated.

A value of 10000 seems a good start. It doesn't seem a reasonable
use case to search for anything that long, and smaller texts shouldn't
cause any freezes. But just in case we want to fine-tune it,
make it a registry key.

GitOrigin-RevId: 8d759e6597ff0df4fa854e77eb3baac9b638ab42