IJPL-163868 Extract FindSettingsBase
Since we can't easily get rid of the public API FindSettings,
the only way to split settings for Find in Files
and Find Usages is to duplicate those settings.
To duplicate them without duplicating the code,
extract the actual implementation into a base class first.
Then we can proceed with creating an extra class for
the Find Usages settings.

GitOrigin-RevId: 517d2e1ae25b588188f3f2a32065f64c7c8de5c6