IJPL-163868 Introduce FindUsagesSettings
To separate Find in Files and Find Usages setting,
create a new interface and its implementation.
Because we have a lot of code relying on the existing
FindSettings interface for Find Usages,
introducing a brand new interface would lead to a lot of code duplication.
Therefore, to avoid breaking compatibility and changing a lot of code,
reuse the existing implementation (extracted into FindSettingsBase)
and make the methods of the new interface have exactly the same
signatures. This way FindUsagesSettingsImpl can reuse existing
implementation while implementing the new interface.

GitOrigin-RevId: be14625f8b6c8777eee7af08a4a846f30e4f85f4