KTIJ-30991 [kotlin] Properly detect and resolve conflicts in Import Optimizer in K2 Mode
The code in `OptimizedImportsBuilder` is mostly a rip-off from
the K1 `OptimizedImportsBuilder` class.

It currently lacks a possibly important optimization - we do not
perform the scopes check to understand if we need to perform
the resolve in the air.

This optimization will be implemented later if deemed necessary.

^KTIJ-30991 Fixed


(cherry picked from commit 50fd5f38605000aaff4eb29aad70604a82a85e4c)

IJ-CR-149236

GitOrigin-RevId: b4355820af690c8cd76a2f8cc2aae2af0b5ab8c9