PY-75909: Expand project view right after new project creation.
Platform expands project view automatically once `generateProject` returns but only if it isn't empty. As we generate things in background (see `.launch`) nothing is expanded.

So we expand it explicitly 2 times: when SDK is generated and when project is generated. At least one of these actions must create files which is required to expand the tree.

This code can't be tested because there is no project view tree in a headless mode.

GitOrigin-RevId: 2390d09619639e8ff8d8906a4e0a284cf24da5ba