WI-76036, WI-76502, WI-77425 lazily search for a `java.applet.Applet` class
I’m not sure why, but during the `PhpGlobalInspectionTestCase` test run, an instance of `RefJavaManagerImpl` is created. In its constructor, the `findClass` method is called, which accesses indexes -> when `ensureUpToDate` is called, all PHP test files are considered as dirty and reindexed, which causes a strange stub behavior that I can’t explain.

GitOrigin-RevId: c40e9f69b4c83349b4ebc9c3758b0c735a41155d