IJPL-164502 Fix restoring maximized state on Sequoia
The cause: setting extendedState to maximized
has the same effect as option-clicking the green button.
Which, on Sequoia, actually un-maximizes the window
if it's already maximized. Whether it's maximized or not
is determined by its size alone. If there's no "normal"
size to restore, then macOS just makes up one,
slightly less than the screen size.

This leads to a bug when the saved size happens to be
the same as the maximum one, which happens quite often.
Then the size is restored, and when extendedState is set,
it makes the frame smaller than it should be.

The fix: if the saved state is maximized, do not restore
the saved size, but instead restore some size that's
smaller than the current screen. For this we introduce
an extra parameter to setDefaultSize(), so we can
specify the exact screen (determined by the saved location).
This way setting extendedState works as expected,
maximizing the frame.

To prevent setExtendedState() from saving this fake
size as the "normal" one, swap restoring the state
and restoring the normal bounds in create():
now we restore the state first, and then apply
the normal bounds, overwriting whatever setExtendedState()
could have saved there.

GitOrigin-RevId: 4f190f6ce5b9d7c5dc52f11abbee4fde53a135b4