[gradle][groovy] IDEA-343916 resolve references in build.gradle to a TOML with custom location of included builds
The changes in this commit allow resolving references to version catalogs in included builds of a composite build if android module provides these version catalogs.

I found two existing approaches to resolve a reverence with a name of a version catalog (e.g. 'myLibs') to a TOML file:
1) Using the data from VersionCatalogsModel that is prepared at sync in `CommonGradleProjectResolverExtension#populateProjectExtraModels`.
Resolve happens in `GradleProjectExtensionContributor`. It works well for simple projects, not composite ones.
This allows navigation to a version catalog with custom location that could be specified in settings.gradle

2) Using the data from `ProjectBuildModel` that is prepared in android module. Resolve happens in `GradleExtensionsContributor`.
It works only if the code in android module correctly discovers version catalogs and provides them.
In comparison to the first approach, the data is compared not while syncing but when it is requested

GitOrigin-RevId: d2859302e5aba22413d3b1da874bff1cd15c46b2