Fix import order for TypeAlias elements
There's an existing test for TypeAlias import ordering, but it's only
passing by accident. Adding a single parameter to the annotation that's
being aliased causes the order to get swapped, which doesn't match the
expected behavior. (I'm assuming the expected behavior is that the
typealias should always be lower on the list; this is based on the
existing test case.)

The existing test case worked because the annotation had a default
constructor with 0 arguments, and the referencing location that needs an
import also has 0 arguments. But any time that argument count doesn't
match, then the order is no longer determined by this weigher.

This fix simply demotes any TypeAlias elements slightly.