IJPL-165109 Fix CachedImageIcon.getScale
Having it hardcoded to 1.0 broke the ScalableIcon
contract: that in order to scale by N times you need to call
scale(N * getScale()), but since getScale() was always 1.0,
it scaled the original icon by N times, not the current one.
So double scaling never worked according to the contract.

Fixing scale() to scale relative to the current size
would be another way to follow the contract,
but it would be more invasive (a lot of code uses
IconUtil.scale) and would still be confusing
(why scale(2.0) returns an icon with a scale of 1.0?).

GitOrigin-RevId: 47a4d2090678d808f00fa352651d2f31f550cd31