IJPL-165109 Fix icon scaling in PopupFactoryImpl
The previous implementation would not take the previous
scaling factor into account. This caused issues such
as CWM-9506: an already scaled icon would make its way there,
and it would be scaled incorrectly.

The correct logic: divide the new size by the old size,
we'll get the "how much bigger do we need it to be?" value.
Multiply it by the current scaling factor according to
the ScalableIcon contract. This is the new scaling factor
that should be passed to scale(). But if it's the same
as the current scaling factor, we don't need to scale.

GitOrigin-RevId: 0e5377f24328fbdff5fb92291736e72ae5cfce8b