Force Skiko to 0.8.17 (#665)
This Skiko version contains an important fix for memory usage. It is
more recent than the version used by CMP 1.7.0 (0.8.15), but the CMP
team assured us it is a binary-compatible change.

As a result you should see greatly decreased memory usage and GC
pressure in most scenarios, except when the ComposePanel size changes.
That is a more complex fix which will require a number of changes across
JBR and Compose, and will come in time.
GitOrigin-RevId: d9e3bc1947760812e66e7932b7f25747632284ef