[RDCT] GTW-2902: Prevent creating editor windows from backend
If an editor is opened with `openMode = NEW_WINDOW`, a new window is created from both frontend and backend. It is unnecessary on the backend for remote clients, it is later ignored and not used in code in any way. This commit prevents the creation of new windows for remote clients.


(cherry picked from commit ae6b7d4d316272568f8465700fcabc97bd10be73)

IJ-CR-148304

GitOrigin-RevId: a740e8081c6c2dc354ad984a740bf73857c2727c