IJPL-164646 Fix coordinate systems in IdeGlassPaneImpl
Looks like some very old bug: the point was converted
into the content pane's coordinate system,
but it was the content pane's parent that was used
to determine the component under the cursor.

This didn't affect anything at first because
the coordinate systems were effectiely the same.
However, when we started using custom frame
decorations, the systems diverged:
the content pane doesn't include the header,
while its parent (the layered pane) does.

Fix by using the parent for both calls.

GitOrigin-RevId: 041dba54d00e803af025052554488f4621ba10f2