IJPL-165416 Add a renderer-style icon to the Test Animator action
A component may be used without actually showing it,
as a "rubber stamp," like tree/table/list cell renderers.
This applies to AnimatedIcon as well, but with an important
caveat: by default, it's only animated when it's showing,
so animation must be started explicitly.

However, if it's started and never stopped, which often happens,
then we'll have a coroutine/memory leak. This can be sort of mitigated
by using Disposable, but disposable Swing components aren't common,
so a lot of usages just ignore it.

The first step to solving this issue is to reproduce it.
We already have the Test Animator action,
so let's add another icon there that demonstrates
the issue.

GitOrigin-RevId: 02d5980fceb3b1cbe9b62e6b1120baee4154e1c5