IJPL-165730 Move logical structure builder creation to a BGT
The way structure view builders are designed, when they're created,
there can be a lot of PSI work. But when the createStructureView
method is invoked, there's a lot of UI creation.

Currently the builders are created on a BGT, which is correct,
and after a recent change, createStructureView is now invoked
on the EDT, which is also correct.

But PhysicalAndLogicalStructureViewBuilder specifically
performed a lot of slow ops stuff in createStructureView.
Let's move that stuff into its constructor instead to fix
those slow ops.


(cherry picked from commit 9eb0470cdd9e59ce39049f6df2696019c03569fd)

IJ-CR-149552

GitOrigin-RevId: 1b8863c788555d20be333dd309bd15deac256a11