IJPL-165416 Fix the coroutine/memory leak in AnimatedIcon
The problem: someone calls resume() when the icon is not showing,
and never shows it (e.g. to use the icon in a renderer).
This causes the animator coroutine to run forever.

The solution: make a special "hidden" animator that doesn't
start a coroutine, but mimics the animator's way of computing
the current frame. When the current frame is requested
on paint, compute it on demand from the current time.
This effectively leads to the same "animation" except
that frames only produced when the component is actually
repainted (which is something the owner should take care of).

The running state of both animators is updated whenever
the showing state changes or the client calls suspend()/resume().
Only one animator is ever running.

Now if the client starts animation and never shows
the component, the "real" animator will never start,
thus not creating a coroutine/memory leak.

GitOrigin-RevId: 37b4724cb62cba2a01f6cbce8f2ae3d88e1e8226