IJPL-165422 Fix Rhizome breaks if exception is thrown under `withKernel`
`withKernel` method attaches current coroutine to core `kernelCoroutineScope`
If current coroutine is finished with an exception, the cancellation is propagated to the parent job.
This cancels `kernelCoroutineScope` and prevents further usage of the scope. The scope is never recreated, so it becomes impossible to use Rhizome after that.

As a solution, `kernelCoroutineScope` should be a supervisor to avoid it being canceled by children.
We can do that by creating `childScope` under the root scope.

GitOrigin-RevId: 54747a102f9fd5d4cc3ca71ae3a793073413b169