Add K2 debugger evaluation tests with compose and parcelize compiler plugins
We have to test whether the compiler plugins correctly work or not for
debugger evaluation. For example, When the debugger evaluates
`Foo(::getString)` for the following code:
```
@Composable
fun Foo(child: @Composable () -> Unit) {
    child()
}

fun getString() = "empty"

fun main() {
..
//Breakpoint!
}
```
it must report a diagnostic error triggered by the compose compiler
plugin checker because it passes a non-composable function for a
parameter with a `@Composable` annotation.

This commit adds a debugger evaluation test to check the above example
reports a diagnostic error correctly.

In addition, the above KT file itself does not have any syntax/semantics
errors, so compiling it should work without an error.  If we run compose
compiler plugin IrGenerationExtension multiple times, it will have some
errors from IR generation for `Foo()`.

To test the registration of compose and parcelize compiler plugins on
IDE and their functionality for the debugger evaluation, this commit
updates `KotlinDescriptorTestCase` to set the test module with the
compiler plugins.

KT-72347