[terminal] IJPL-165429 IJPL-159646 Fix shell editor buffer reporting for ZSH
Previous fix works for bash, but breaks zsh. This happens because zsh does not support prompt_shown event, which is required for this to work.
Now we report buffer on both prompt shown and prompt state updated, which is effectively when the prompt is shown.
Fix shell editor buffer reporting for ZSH

Previous fix worked for bash but broke zsh due to its lack of support for the prompt_shown event. Now, the buffer is reported both on prompt shown and prompt state updated, ensuring compatibility with zsh.

GitOrigin-RevId: 43a8d35e12c74d4db0f59c55f5ae52f7d38283bb