IJPL-158493 Redesign tree models: leaf state, visitor, presentations
Move leaf state computation to the domain model.
It now returns just a boolean.
It was very hard to do it in the view model because
the children had to be either requested twice,
or cached in a very roundabout way.

The same with the comparator. Now it's the responsibility
of the domain model to provide data in the correct order.

Drop presentation internals that should be of no interest
to domain model providers.

Drop legacy visitors support on the view model level.
Keep it on the swing model level for now.

Add the parent to the tree node view model,
and together with the dropped visitor support
this finally allows us to mostly get rid of the Swing
dependencies there. Except icons.

GitOrigin-RevId: a2e4c774b29cc95c0ecc59b0951b1ce6c04de582