[kotlin] find usages: detect enum's constructor usages
enum constants are represented as a light version of PsiEnumConstants, but they provide no references. Thus they are not processed by java's ConstructorReferencesSearchHelper (which is good, otherwise it would be complicated to provide usages only once in cases, where there is no java). So additional searcher is provided for kotlin.

^KTIJ-31884 fixed

GitOrigin-RevId: b7bc32801ff907806a1d971a082633aad2bbe794