IJPL-166285 Fix accessing unregistered RPC breaks following RPC calls
If RPC is not followed, backend throws an error `No remote API found for InstanceId` in `RemoteApiRegistry.resolve`.
`RpcExecutor`, however, expects this method to return `null` if service was not found, not an exception.

Therefore, the exception is propagated upward which breaks the kernel and prevents further RPC calls from executing.

We should adhere to the API requirements and return `null` if service was not found instead of throwing an error.

GitOrigin-RevId: 1388465c6cbfcef6467b03fa420f3b575db2b58a