IJPL-158493 Implement TreeViewModel.awaitUpdates
Now that we finally have all the updates in one place,
we can number them and track them.
No way to track every single one of them,
of course, as they're asynchronous,
but we can wait until it's
"at least up to date as the last update,"
which is good enough for most purposes.

GitOrigin-RevId: 199d56b02fb883d48369952694ff1ce70ddc41c1