IJPL-165735 Fix the JTree.expandedState leak in Tree
JTree.setModel() is the only place that modifies expandedState
that we haven't overridden when implementing the bulk expand API.
On one hand it's just one tree path, but on the other hand
that tree path can contain references to a huge tree structure,
thus creating a memory leak even after clearToggledPaths()
was explicitly called.

The fix is two-fold: clear expandedState on model change
and call super.clearToggledPaths() in clearToggledPaths()
explicitly. This ensures both that expandedState remains empty,
and in case we've missed a similar bug elsewhere,
it at least guaranteed that both JTree and Tree expanded states
are cleared when the client does so explicitly.

GitOrigin-RevId: 246fb1fddde5a4bfacfe7400af90a008f2903c30