[remote dev] bundle keymap plugins with frontend (JetBrains Client) distributions (RDCT-1702)
Now frontend distribution bundles the same keymap plugins as the corresponding full IDE, and build tests verify that. This is needed to ensure that compatible versions of these plugins are always installed, and there is no need to update them manually if an EAP version of the IDE was used before (see GTW-9593).

GitOrigin-RevId: fb59cafc43146aec7aae016993a04f79f3366700