IJPL-166992 When tab is restored from editor Hide button does nothing
After the placeholder is set to the content, we can tell the explicitly requested focus target from the calculated one by the policy.

Problems happen when that placeholder is set as explicitly requested by mistake. It is then kept as such after the tab is restored. Then it ruins the data-context for a toolwindow toolbar.

GitOrigin-RevId: fa503c088aab0720dd74052657a7b83612ee4180