[exception analyzer] Add a registry key to avoid showing notification about successfully reporting an error
Successful error reporting is a normal situation, and it doesn't deserve any notification. At the same time, some find it useful to open a report after submitting and to use the link provided in the notification. Thus a registry key to be able to utilize both approaches. Initially set to "show notification" to keep the current behavior.

GitOrigin-RevId: 194232347dadf1918d80ff543cfb7dfc4f0ccffa