IJPL-158493 Include expanded state in the tree view model
In Swing it used to be a part of the tree state,
but here for UDF we want everything to be a part of
the view model (the single source of truth).

To avoid spamming a lot of flows inside every node
view model, replace the existing presentation with
a more generic state flow, which currently includes
the presentation and the expanded state
(the selected state to be included shortly).

Rename the Swing model's state (published, etc.)
to lifecycle to avoid confusion with the view model's state.

Take care not to update the view model state
from Swing if the change comes from the view model
itself. This is because such a change may be outdated,
there may be newer changes coming from the view model,
and we don't want to overwrite those with the outdated
change we're currently applying. So only propagate
Swing state changes to the view model if they originate
from Swing itself (e.g., mouse clicks, key presses,
or explicit Swing API calls).

GitOrigin-RevId: f0f1265e384bd97ea4d32effa24a96791b50669a