[debugger-dfa] Forcefully rewrite variable constant value in debugger
If any previous knowledge available for variable (e.g., from NotNull or Range annotation) contradicts the value observed in the debugger, we should not refuse to update it. The debugger value should have priority, as it's what we really observe, even if the contract is against it.
Fixes IDEA-361617 Conflicts between annotations and real value in debugging

GitOrigin-RevId: 34e6454a60fb9d381b58ee5ca84b1606e0c4afe2