IJPL-160390 Expand `SuspenderListener` API
To support suspendable progresses I need an ability to map `ProgressSuspender` to `ProgressIndicator` when a suspender appears.

Before, it was only possible to track new `ProgressSuspender` creation.
If one suspender is attached to several indicators - `suspendableProgressAppeared` won't be called.
Now, there is a second method that is called for every new progress attached to the suspender.

Also, `ProgressSuspender` can be closed at any point in time, and it'd mean that the progress is no longer suspender.
The listener should notify about this too

GitOrigin-RevId: f1b66f4f96d85e013df799fa7b7e52896fd89d3a