IJPL-158493 Implement selection in the tree view model
Along with the expanded state, this more or less implements
the full state of a tree in the view model.

There are some things to add, like support for centering
selection, but as far as the state of the tree+model combination
goes, we now have everything, or almost everything
(we probably need to add isRootVisible too).

GitOrigin-RevId: 34d64e56e3bade0c04bd2f73955bcf3999ff62c2