IJPL-158493 Introduce TreeSwingModelListener
Two reasons for introducing.

1. Using TreeSelectionListener for this purpose
was more like abusing it. Its old school
interface with a lot of parameters didn't suit the task
well enough, and most of those were unused,
with the list of paths having different semantics.

2. We now need to add more methods to it
for scrolling to selection. The existing interfaces
don't support this, so we need a new one anyway,
so might as well just use it for everything.

GitOrigin-RevId: 5d7be2f7788da602e8a58425327592f49e2993ec