IJPL-158493 Implement tree view model scroll requests
With the Swing part it's tricky as usual:
we must wait for the node to be published
before scrolling. Can't use the existing awaitNode()
implementation in the visitor because that one
relies on running within the node's scope
to ensure cancellation on node cancellation.

Luckily, we can just store the reference and scroll
whenever the node is published. This creates a risk
of a leak, though, but let's think about it later.

GitOrigin-RevId: ccf99a5cc2c1fee095b95dea616c1d3cd782f360