IJPL-158493 Drop the ide.tree.bulk.expand.api key
The new implementation is stable for a few releases already,
it works, nobody uses the legacy (Swing) one.
And it's not easy to integrate the old one with the new view models
because it's tricky to sync the expanded state of the model
with that of the tree without modifying the tree side implementation.
And we can only do that with our own, not with the Swing one.

Drop some dead code while we're at it. In theory,
it would seem that we could also drop all those expandImpl != null
checks, but it's not that simple: JTree invokes a lot of that stuff
from the constructor, so it CAN be null. Ideally, we should go over
those cases and handle them one-by-one instead of just defaulting
to the legacy implementation there (which doesn't make sense anyway),
but it's a risky job and definitely out of scope of the current task.

GitOrigin-RevId: a372d20348f3b4f77ec9cdba4a193d966c970327