IJPL-158493 Fix tree/model expand/collapse sync
The new view model and JTree have different approaches
to expand/collapse. The view model is "dumber":
when a node is collapsed, it's collapsed and nothing more.
The tree is trying to be "smart" about it: when a node is
collapsed, its ancestors are expanded to "show" that
the node has indeed been collapsed. Whether it makes
sense or not, is debatable, but as we must preserve the existing
contracts of the tree, let's just not mess with the already working
code and fix the sync part instead: if the expanded state comes
from the view model, and the node is invisible, just apply the new
state to it and don't do anything more.

This fixes a bug when collapsing some nodes in a random order
would lead to expanding some of them afterward.

To save some memory on leafs, when marking a collapsed
node as such, remove its entry if it's a leaf.

GitOrigin-RevId: 56b6582c6a6201cb26dfc373da306103c6d80618