[python] Untie PyNarrowedType from PyCollectionType
PyNarrowedType is not a proper generic type, in the sense of a type extending
typing.Generic and having explicit type parameters. It's a parameterized
"special form", such as typing.Literal, that can also be treated as
PyClassType(bool). Representing it as "bool[T1, T2, T3]" (even
though it can contain only one type) is confusing and potentially
error-prone.

Also, extracting it from a type hint shouldn't rely on
PyTypingTypeProvider.getParameterizedType.
The latter should really work only for types extending typing.Generic,
which is required for addressing PY-77074.

GitOrigin-RevId: 931eb73590b9b0b3a96cd27f9e5201d079be5228