[kotlin] Fix checking a method to have a Java origin in overridden symbols
If a method overrides some Java method and that method is not overridden with anything else, it means that this method originates from Java.
In comparison to the previous wrong implementation, for the case `class ClassInheritingBoth: JavaName, KotlinName` the current implementation works regardless of the order of implemented interfaces `JavaName`, `KotlinName`.

^KTIJ-30819 Fixed

GitOrigin-RevId: eb262ccbf0c4c38248a260130d58316fa7a8bb65