IJPL-171464 Calculate column widths even if find.usages.disable.smart.size is on
When we added this key, we completely disabled all resizing logic
when Show Usages appears. This works fine for the size of the popup
itself, but the column sizes are now all wrong.

Adding this one line seems to fix the column sizes without affecting
the size of the popup.


(cherry picked from commit dd0659caa779f533770451a31ec3807394ab2c66)

IJ-CR-149328

GitOrigin-RevId: fd9ea59a1b841ba2ba23f782733c6d8dee574fec