IJPL-158493 More tree debugging
Add a debug message about propagating the expanded state
from the model. This helps with investigating
"Why has this thing collapsed?" issues.

Add the expanded state to the view model's toString().
This helps to track this state in debug messages.
Not worth including the entire state there, as the presentation
can be quite huge (with all those icon toString()s).

GitOrigin-RevId: 70d027b21a1e0fc20e57acd5165a037d4f737e4c