IJPL-171351 Add the final popup screen bounds debug logging
When the popup is finally shown, the log message
contains the coordinates of the popup and the ID
of the target screen. Add the bounds of the target
screen as well. They may or may not appear in earlier
messages, but it's very useful to have the end result
in one place (for example, to reason about why
it was / wasn't necessary to change the popup
bounds to fit it into the screen).


(cherry picked from commit d1f4fea44c4749589fccbb914afcc450d48df511)

IJ-CR-149457

GitOrigin-RevId: 820ba1ff87ab632dd2e206eac1751683dc25c622