IJPL-158493 Make Tree.expandImpl @NotNull
It's really confusing to have all those null checks for
something that is not null now.

To address the case of accessing this thing from
the superclass constructor, a separate flag is introduced.
All cases of access were followed through and guarded
by this flag. This includes:

1. firePropertyChange() from setModel():

1.1. Expanding the root path. This is now ignored here
and later handled in the new ExpandImpl constructor,
where this state is initialized and the super() state is cleared
to avoid leaks.

1.2. Setting the cached presentation. It can't be set yet
at this point, so this part we just ignore.

2. isExpanded() and getExpandedDescendants().
These things are used in
BasicTreeUI.updateLayoutCacheExpandedNodesIfNecessary
to copy the "root expanded" state to the tree layout cache.
These now delegate to super() to ensure this state is properly handled.

3. clearToggledPath(): called from setModel() and is effectively
a no-op at this point because there's no previous state to clear.

4. The tree model listener now simply ignores
all events before its delegate is initialized.
Ideally, a model shouldn't fire events at all while
we're still in setModel() (called from the super constructor),
but some models do it anyway, like AsyncTreeModel,
when getRoot and getChild are called.
It doesn't make sense to try to handle those events,
though, as the expanded state is initialized later anyway,
and everything else like the cached presentation
is irrelevant at that point. If something comes up,
we can initialize that later too, whatever it is.

Let's hope we haven't missed anything, but these exceptions
are pretty serious, it'll be easy to detect them even if we
did miss something.

GitOrigin-RevId: fbe12424f3893e679cabb63e6cde66c2ef7b419d