IJPL-171570 Wait for the UI scale to init before initializing LaF
The issue described in the ticket seems to stem from the fact
that on macOS there's no guarantee that the UI scale is initialized
even as late as when initLafManagerAndCss is called.

The natural place to ensure it seems to be initLafJob,
as initLafManagerAndCss waits for it to complete the first thing,
and it seems natural to group these things together: the UI scale
init, base LaF init and configureCssUiDefaults. Any place
that relies on initLafJob to complete (like initLafManagerAndCss)
is very likely to also expect the UI scale to be initialized: LaF
doesn't make sense without it.

On non-macOS the added line is effectively a no-op.
Or, to be more precise, it makes the next line a no-op,
as initUiScale is essentially the same then as initBaseLafJob,
and there's no harm in joining the same job twice.

GitOrigin-RevId: e1f4a512cacdf1ac7d93967819980861d47ffa08