[kotlin] implement `KotlinProjectStructureProvider` without using `ModuleInfo`
The implementation is based on a workspace model. It gets `WorkspaceEntity` for a `VirtualFile` and, depending on it, creates the corresponding `KaModule`. If there are multiple candidates, it chooses the one on which the use-site module depends, considering the classpath priority.

The implementations of `KaModule` are created from scratch to avoid using `ModuleInfo` inside. Instead, they store the `SymbolicEntityId` of the module.

During a transitional period, it's possible to switch to the old implementation by using the `kotlin.use.new.project.structure.provider` registry key.

Semantic changes compared to the previous `ModuleInfo`-based implementation:
- `KaLibraryModule` implementation has no dependencies. In the previous version, dependencies were computed based on the dependencies which dependent modules had. That list was incomplete and used for K1-resolution. K2 resolution does not need that info, so libraries do not have any dependencies now.
- `KaLibraryModule` is no longer split into different klib based libraries. Instead, it has a single `KaLibraryModule` module as in the project model.
- `KaLibraryModule` modules are now not deduplicated based on the same source roots. This seems to cause no performance degradation and simplifies the logic.

^KTIJ-31422 fixed

GitOrigin-RevId: 9ba6bcb6fb44cd369740577b9377096e6a80759b