Add extension point for custom compiler plugin configuration
When we register a compiler plugin via
KotlinBundledFirCompilerPluginProvider, we cannot set the compiler
plugin configuration. For some compiler plugins, setting the wrong
compiler plugin configuration is critical. For example, the parcelize
compiler plugin has a plugin option `USE_FIR`. If we does not set it,
it chooses the default value i.e., `false` and causes issues on IR
generation.
This commit add `KotlinFirCompilerPluginConfigurationForIdeProvider`
that allows us to set the compiler plugin configuration. We add a
default custom compiler plugin configuration provider for the parcelize
compiler plugin (`DefaultParcelizeFirCompilerPluginConfigurationForIdeProvider`)
to set `USE_FIR` option correctly.

^KTIJ-31698

GitOrigin-RevId: ca44710144128c793afb99f98c4b4891aa819fcb