use LF line endings for all text files in intellij-community by default (IJI-526)
Before, it was recommended to set `core.autocrlf` Git option to `input` before checking out the repository, and most of the files in the repository use LF line endings. However, it was easier to forget changing the option. In that case you will get CRLF line endings on Windows, and it leads to problems in some cases (see IDEA-297137, for example). Now usage of LF line endings in the working tree is forced via the .gitattributes file. If it's necessary to use CRLF line endings in a specific file, it must be specified explicitly by via a .gitattributes file.

GitOrigin-RevId: 7ac559d8f3ec32f2a37c01937d233e26f821c664